requirements.txt
--index-url https://pypi.org/simple

wheel
typing-extensions
pysigrok-libsigrokdecode
